Version 4.0.2308.2201
https://gurux.fi/release/guruxdlmsami4-4023082201

+++ Gurux.DLMS.AMI.Agent.Worker/GXAgentWorker.cs
@@ -716,7 +716,7 @@ namespace Gurux.DLMS.AMI.Agent.Worker
                     int deviceAddress;
                     if (settings.HDLCAddressing == (int)ManufacturerSettings.HDLCAddressType.SerialNumber)
                     {
-                        deviceAddress = GXDLMSClient.GetServerAddressFromSerialNumber((int) settings.PhysicalAddress,
+                        deviceAddress = GXDLMSClient.GetServerAddressFromSerialNumber(settings.PhysicalAddress,
                             settings.LogicalAddress,
                             settings.SerialNumberFormula);
                     }
